+++ gfx/resources/OtShaderProgram.h
@@ -67,7 +67,7 @@ public:
 	// set parameters for the next run
 	inline void setState(uint64_t state) { bgfx::setState(state); }
 	inline void setTransform(const glm::mat4& transform) { bgfx::setTransform(glm::value_ptr(transform)); }
-	inline void setTransforms(const glm::mat4* transforms, uint16_t count) { bgfx::setTransform(transforms, count); }
+	inline void setTransforms(const glm::mat4* transforms, size_t count) { bgfx::setTransform(transforms, static_cast<uint16_t>(count)); }
 
 	// run program in the specified view on the GPU
 	void submit(bgfx::ViewId view);
